+++ src/aircraft.c
@@ -11,12 +11,17 @@
 #include <math.h>
 #include <stdlib.h> 
 
-void initAircraft(AircraftState *aircraft) {
+void getEmptyMassAndMaxFuel(AircraftData *aircraftData, float *emptyMass, float *maxFuel){
+    *emptyMass = (float)aircraftData->mass; // kg
+    *maxFuel = (float)aircraftData->fuelCapacity; // kg
+}
+
+void initAircraft(AircraftState *aircraft, AircraftData *data) {
     // Initialize the aircraft's state with zero values or default values
     aircraft->x = 0.0f; // Set initial x position to 0
     aircraft->y = 500.0f;  // Start at an altitude of 500 units
     aircraft->z = 0.0f; // Set initial z position to 0
-    aircraft->vx = 50.0f;  // Set an initial forward velocity of 50 units
+    aircraft->vx = 100.0f;  // Set an initial forward velocity of 50 units
     aircraft->vy = 0.0f; // Set initial y velocity to 0
     aircraft->vz = 0.0f; // Set initial z velocity to 0
     aircraft->yaw = 0.0f;   // Set initial yaw to 0 (facing straight)
@@ -24,6 +29,11 @@ void initAircraft(AircraftState *aircraft) {
     aircraft->roll = 0.0f;  // Set initial roll to 0 (level flight)
     aircraft->AoA = 0.0f;  // Set initial Angle of Attack to 0
     aircraft->hasAfterburner = false; // Set afterburner to false by default
+
+    // fuel, mass
+    getEmptyMassAndMaxFuel(data, &aircraft->currentMass, &aircraft->fuel);
+
+    aircraft->currentMass += aircraft->fuel; // Add fuel to the current mass
 
     // Initialize controls
     controlsInit(); // Call the function to initialize controls